% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CopyImage}
\alias{CopyImage}
\title{Copies raw airphotos to 0_raw local directory for further processing.}
\usage{
CopyImage(x, outdir = "0_raw")
}
\arguments{
\item{x}{full path to images (full path).}

\item{outdir}{local directory to store image copy}
}
\description{
Copies raw airphotos to 0_raw local directory for further processing.
}
