% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{CorrectLayers}
\alias{CorrectLayers}
\title{Adjusts imagery to reduce artifacts associated with flightlines, and phenology inconsistencies.}
\usage{
CorrectLayers(y, x, maxsample = 1e+05, sampdens = 50)
}
\arguments{
\item{y}{raster object created from seamlines, containing two layers, one describing distance from seamline, the other indicating image date.}

\item{x}{raster to be corrected}

\item{maxsample}{maximum pixels to be sampled for corrective model}

\item{sampdens}{sampling density to be sampled for corrective model.}
}
\value{
multiband raster containing layers adjusted according to information associated with image acquisition date, and also distance to seamline (note: flightline distance would be better, but not currently available)
}
\description{
Corrects image artifacts based on seamlines, image dates in airphoto metadata polygons.  Alternate sources of imagery corrections could be used.
}
