% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetAFFT}
\alias{GetAFFT}
\title{Generate metrics aggregating information from full-res images to coarser resolution.}
\usage{
GetAFFT(
  filelist,
  zradii = c(0.75, 1.25, 2.5, 5, 10, 60),
  outres = 30,
  overwrite = T,
  ncpu = 7
)
}
\arguments{
\item{zradii}{radii for summarizing fft-related statistics.}

\item{outres}{resolution of output raster}

\item{r}{tif image name}
}
\value{
raster containing multiple bands, describing texture at scales described over scales indicated by zradii, as well as an array of non-textural summary statistics.
}
\description{
Generate metrics aggregating information from full-res images to coarser resolution.
}
