% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetFootprints}
\alias{GetFootprints}
\title{Grabs extents from all airphotos imglist.}
\usage{
GetFootprints(imglist, outcrs = NULL, return.as.list = F, quiet = T)
}
\arguments{
\item{imglist}{list of airphotos (text, full paths, with names)}

\item{outcrs}{projection object if reprojection is needed}

\item{return.as.list}{sets return to one airphoto extent (terra vect format) per object in a list (as opposed to one vect object)}

\item{quiet}{set to F if more intermediate messaging is wanted.}
}
\value{
Terra vect object containing airphoto extents, either as a single object, or as a list of extents.
}
\description{
This function is useful for imputation mapping from models built with the yaImpute package.
}
