% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetFullResImageList}
\alias{GetFullResImageList}
\title{internal.}
\usage{
GetFullResImageList(
  fn,
  subset1 = c("r", "g", "n", "ndvi", "ndng", "ndgr", "bri"),
  rawpath = "0_raw",
  indpath = "1_intermediate",
  indfolders = c("ndvi", "ndng", "ndgr", "bri")
)
}
\description{
internal.
}
