% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetMetrics}
\alias{GetMetrics}
\title{Calculates textural summaries of input images at an aggregated scale.}
\usage{
GetMetrics(
  rasterfile,
  outpath1 = "1_intermediate",
  outpath2 = "2_aggregated",
  outres = 30,
  ncpus = 3,
  zradii = c(3, 6, 10, 20, 40, 55)
)
}
\arguments{
\item{rasterfile}{character vector with tif image names. If full path not included, GetMetrics will assume that files are in 0_raw folder.}

\item{outpath1}{file path pointer where R will store intermediate (full resolution) image tiles, with calculated indices as well as original bands (although not blue currently)}

\item{outpath2}{file path pointer where R will store aggregated image tiles (reduced resolution), texture summaries across all bands and indices.}

\item{outres}{resolution of output texture summaries}

\item{ncpus}{specify number of cpus to use for parallel processing.}

\item{zradii}{numeric vector specifying frequency-length bins for summarizing fft spectrum. (maximum value of this should be larger than diagonal of output pixel size)}
}
\value{
multiband raster
}
\description{
This function is useful for imputation mapping from models built with the yaImpute package.
}
\seealso{
\code{\link{MakeDonut}}
}
