% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetMetrics1}
\alias{GetMetrics1}
\title{Calculates indices from RBGN. Resolution unchanged. internal for GetBandIndices}
\usage{
GetMetrics1(
  rasterfile,
  outpath = "1_intermediate",
  fl = indexFuns,
  parallel = F,
  quiet = T
)
}
\arguments{
\item{rasterfile}{character vector with tif image names. If full path not included, GetMetrics will assume that files are in 0_raw folder.}

\item{outpath}{file path pointer where R will store intermediate (full resolution) image tiles, with calculated indices as well as original bands (although not blue currently)}
}
\value{
multiband raster with ndvi, ndgr, ndng, brightness.
}
\description{
Copies source airphotos to intermediate directory, calculates extra indices to add to image. Resolution unchanged.
}
