% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{data}
\name{indexFuns}
\alias{indexFuns}
\title{functions for tallying indices on raw airphotos. pass this object (or revised indices) to cluster with sfExport.}
\format{
An object of class \code{list} of length 4.
}
\usage{
indexFuns
}
\description{
List of functions for tallying indices on raw airphotos. pass this object (or revised indices) to cluster with sfExport.
}
\keyword{datasets}
