% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeDonut}
\alias{MakeDonut}
\title{Used within GetMetrics.}
\usage{
MakeDonut(zr, res1, fact1, return.rast = T, inproj)
}
\arguments{
\item{zr}{Radii for nested circles}

\item{res1}{output resolution (full size indicated by the output matrix)}

\item{fact1}{resolution of imagery to be summarized}
}
\value{
matrix with integers that is used for extracting zonal summaries of fft spectrum. Donut values indicate scale of variation in meters.
}
\description{
Creates a matrix whose outer size matches the dimension of the image to be summarized (in this case, an airphoto image covering exactly one landsat pixel)
}
\seealso{
\code{\link{GetMetrics}}
}
