% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeDonut2}
\alias{MakeDonut2}
\title{Used within GetMetrics.}
\usage{
MakeDonut2(zr, res1, inrast)
}
\arguments{
\item{zr}{Radii for nested circles}

\item{res1}{output resolution (full size indicated by the output matrix)}

\item{inrast}{input raster of imagery to be summarized (used to get resolution)}
}
\value{
matrix with integers that is used for extracting zonal summaries of fft spectrum. Donut values indicate scale of variation in meters.
}
\description{
Creates a matrix whose outer size matches the dimension of the image to be summarized (in this case, an airphoto image covering exactly one landsat pixel)
}
\seealso{
\code{\link{GetMetrics}}
}
