% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakePCA}
\alias{MakePCA}
\title{Perform principal components analysis to help reduce dimensionality of a multiband image.}
\usage{
MakePCA(r, sampdens = 50, maxsample = 1e+05)
}
\arguments{
\item{r}{multiband rast image.}

\item{sampdens}{sampling density for building PCA}

\item{maxsample}{maximum number of samples to draw from image.}
}
\value{
multiband raster containing pca axis scores.
}
\description{
Creates a matrix whose outer size matches the dimension of the image to be summarized (in this case, an airphoto image covering exactly one landsat pixel)
}
