% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MergeAggregatedAirphotos}
\alias{MergeAggregatedAirphotos}
\title{Used after GetMetrics.}
\usage{
MergeAggregatedAirphotos(tiflist, outpath2 = outpath2)
}
\arguments{
\item{tiflist}{list of files to merge}

\item{outpath2}{file path to folder where output of this function should be stored (usually same directory as tiflist).}
}
\value{
list of images. Mostly called to generate images that are easy to read in and work with later.
}
\description{
merges tiles in tiflist, one tif per band in the tiles.
}
