% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MergeAggregatedAirphotos2}
\alias{MergeAggregatedAirphotos2}
\title{Used after GetMetrics.}
\usage{
MergeAggregatedAirphotos2(tiflist, filename, outpath = "2_aggregated")
}
\arguments{
\item{tiflist}{list of files to merge}

\item{filename}{name of output file}

\item{outpath2}{file path to folder where output of this function should be stored (usually same directory as tiflist).}
}
\value{
output image path. Mostly called to generate images that are easy to read in and work with later.
}
\description{
merges tiles in tiflist, one tif per band in the tiles.
}
