% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ReprojectAFFT}
\alias{ReprojectAFFT}
\title{Reprojects AFFT quarter-quads to match a snap raster.}
\usage{
ReprojectAFFT(fl, snap, outpath, ncpu = 24, p = "D:/LocalNaip/progress.txt")
}
\arguments{
\item{fl}{character vector with full paths to source images named with the image file name with no path..}

\item{snap}{character string, with full path to snap raster image, ideally local.}

\item{outpath}{character string with full path to output folder.}

\item{ncpu}{number of cpus to use in cluster.}

\item{p}{character string pointing to a text file where progress will be reported.}
}
\value{
list of file paths to projected images.
}
\description{
Reprojects AFFT quarter-quads to match a snap raster.
}
