% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ScoreArtifacts}
\alias{ScoreArtifacts}
\title{Extracts fourier summary of image, returns estimate of how much image variability is at a pixel-to-pixel scale (a.k.a. noise)}
\usage{
ScoreArtifacts(y, x, maxsample = 1e+05, sampdens = 50)
}
\arguments{
\item{y}{rast object containing layers that may indicate artifacts in teh imagery (e.g., flightlines, or phenology blocks)}

\item{x}{rast object}

\item{maxsample}{maximum pixels to sample}

\item{sampdens}{density to sample pixels (up to maxsample)}
}
\value{
list of linear models describing correlations between x and y.
}
\description{
Extracts fourier summary of image, returns estimate of how much image variability is at a pixel-to-pixel scale (a.k.a. noise)
}
