% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ScoreNoise}
\alias{ScoreNoise}
\title{Extracts fourier summary of image, returns estimate of how much image variability is at a pixel-to-pixel scale (a.k.a. noise)}
\usage{
ScoreNoise(
  y,
  f1 = 256,
  f2 = 64,
  q = T,
  scorefun = function(x, q = c(0.05, 0.1, 0.5, 0.9, 0.95)) {
     quantile(x, q, na.rm =
    T)
 },
  return.rast = F
)
}
\arguments{
\item{y}{rast object}

\item{f1}{resolution to extract sub-images for fourier analysis.  256 currently used for efficiency}

\item{f2}{secondary resolution for summarizing fourier spectrum. Set to 64 for computational efficiency.}

\item{q}{if set to T, R will print out images as it scores noise for the large image tiles.}

\item{scorefun}{specifies how variability in noise scores for tiled fft summaries will be stored}

\item{return.rast}{if set to T, function returns the image of noise scores with a large pixel size (aggregate of original image to to f1)}
}
\value{
statistics describing the proportion of image variation at very small scales.
}
\description{
Extracts fourier summary of image, returns estimate of how much image variability is at a pixel-to-pixel scale (a.k.a. noise)
}
