% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ViewStackRGB}
\alias{ViewStackRGB}
\title{Useful for viewing multi-band images in rgb space.  Groups bands 1:3, 4:6, ... through the number of layers in the input raster.}
\usage{
ViewStackRGB(x)
}
\arguments{
\item{x}{rast object}
}
\value{
returns NULL.  Used for plotting
}
\description{
Extracts fourier summary of image, returns estimate of how much image variability is at a pixel-to-pixel scale (a.k.a. noise)
}
