% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ViewStackRGB2}
\alias{ViewStackRGB2}
\title{Useful for viewing multi-band images in rgb space - handles two images.}
\usage{
ViewStackRGB2(x, y)
}
\arguments{
\item{x}{rast object}

\item{y}{rast object, same dimensions as x.}
}
\value{
returns NULL.  Used for plotting
}
\description{
Useful for viewing multi-band images in rgb space.  Groups bands 1:3, 4:6, ... through the number of layers in the input raster.
}
