% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ViewStackSingleBandStretch}
\alias{ViewStackSingleBandStretch}
\title{Useful for viewing multi-band images in rgb space}
\usage{
ViewStackSingleBandStretch(x, ext = NULL)
}
\arguments{
\item{x}{rast object}

\item{ext}{ext object}
}
\value{
returns NULL.  Used for plotting
}
\description{
Useful for viewing multi-band images in rgb space.  Groups bands 1:3, 4:6, ... through the number of layers in the input raster.
}
