% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GetAFFT}
\alias{GetAFFT}
\title{Generate metrics aggregating information from full-res images to coarser resolution.}
\usage{
GetAFFT(r, zradii = c(2, 6, 56), outres = 30)
}
\arguments{
\item{r}{tif image name}

\item{zradii}{radii for summarizing fft-related statistics.}

\item{outres}{resolution of output raster}
}
\value{
raster containing multiple bands, describing texture at scales described over scales indicated by zradii, as well as an array of non-textural summary statistics.
}
\description{
Generate metrics aggregating information from full-res images to coarser resolution.
}
